minor traits refactoring
