+++ compiler/src/ast/impl_ast.rs
@@ -42,10 +42,10 @@ use super::ast::{
     UnresolvedIdentifier, UserDefinedTypeNode, VariableDeclarationNode, WhileLoopStatementNode,
 };
 use super::iterators::SymbolSeparatedSequenceIterator;
-use crate::ast::ast::ErrornousNode;
 use crate::ast::ast::MissingTokenNode;
-use crate::ast::ast::Node;
 use crate::ast::ast::SkippedTokensNode;
+use crate::ast::traits::ErrornousNode;
+use crate::ast::traits::Node;
 use crate::code::JarvilCodeHandler;
 use crate::core::string_interner::{Interner, StrId};
 use crate::lexer::token::{BinaryOperatorKind, Token, UnaryOperatorKind};
