+++ compiler/src/ast/mod.rs
@@ -4,4 +4,5 @@ pub mod ast;
 pub mod impl_ast;
 pub mod iterators;
 pub mod print;
+pub mod traits;
 pub mod walk;
