+++ compiler/src/parser/type_checker.rs
@@ -54,11 +54,12 @@ use crate::{
             ComparisonNode, CoreAssignmentNode, CoreAtomNode, CoreAtomStartNode,
             CoreAtomicExpressionNode, CoreExpressionNode, CoreRVariableDeclarationNode,
             CoreSelfKeywordNode, CoreStatementIndentWrapperNode, CoreStatementNode, CoreTokenNode,
-            CoreTypeDeclarationNode, CoreUnaryExpressionNode, ExpressionNode, Node,
+            CoreTypeDeclarationNode, CoreUnaryExpressionNode, ExpressionNode,
             OnlyUnaryExpressionNode, RAssignmentNode, ReturnStatementNode, StatementNode,
             SymbolSeparatedSequenceNode, TokenNode, TypeExpressionNode, UnaryExpressionNode,
             VariableDeclarationNode,
         },
+        traits::Node,
         walk::Visitor,
     },
     constants::common::{BOOL, FLOAT, INT, STRING},
